Fix some minor bugs.

+++ misc/display.py
@@ -126,7 +126,8 @@ def showFrame(frame, del_all=False):
         f = datahandler.ClFits(file)
         if (f.mef==True):
                 # Multi-Extension FITS files
-                if (f.isDark() or f.isMasterDark() or f.isMasterDarkModel()):
+                if (f.isDark() or f.isMasterDark() or f.isMasterDarkModel()
+                    and f.getInstrument()=='hawki'):
                     # (Hawki) Dark files don't have WCS information required by ds9 mosaicimage
                     if delete_all: os.system(("%s/xpaset -p ds9 frame delete all" % (ds9_path)))
                     os.system(("%s/xpaset -p ds9 frame new" % ds9_path))
